+++ packages/dql/src/ObjectProviders/DoctrineOrmEntityProvider.php
@@ -8,9 +8,7 @@ use Doctrine\ORM\EntityManager;
 use Doctrine\ORM\QueryBuilder;
 use EDT\DqlQuerying\Contracts\ClauseInterface;
 use EDT\DqlQuerying\Contracts\MappingException;
-use EDT\DqlQuerying\Utilities\JoinFinder;
 use EDT\DqlQuerying\Utilities\QueryBuilderPreparer;
-use EDT\DqlQuerying\Utilities\QueryGenerator;
 use EDT\DqlQuerying\Contracts\OrderByInterface;
 use EDT\Querying\Pagination\OffsetPagination;
 use EDT\Querying\Contracts\ObjectProviderInterface;
@@ -25,23 +23,19 @@ use EDT\Querying\EntityProviders\OffsetPaginatingEntityProviderInterface;
 class DoctrineOrmEntityProvider implements ObjectProviderInterface, OffsetPaginatingEntityProviderInterface
 {
     /**
-     * @var QueryGenerator
-     */
-    private $queryGenerator;
-
-    /**
      * @var QueryBuilderPreparer
      */
     private $builderPreparer;
 
     /**
-     * @param class-string<TEntity> $className
+     * @var EntityManager
      */
-    public function __construct(string $className, EntityManager $entityManager)
+    private $entityManager;
+
+    public function __construct(EntityManager $entityManager, QueryBuilderPreparer $builderPreparer)
     {
-        $metadataFactory = $entityManager->getMetadataFactory();
-        $this->queryGenerator = new QueryGenerator($entityManager);
-        $this->builderPreparer = new QueryBuilderPreparer($className, $metadataFactory, new JoinFinder($metadataFactory));
+        $this->builderPreparer = $builderPreparer;
+        $this->entityManager = $entityManager;
     }
 
     /**
@@ -93,12 +87,21 @@ class DoctrineOrmEntityProvider implements ObjectProviderInterface, OffsetPagina
         int $offset = 0,
         int $limit = null
     ): QueryBuilder {
-        return $this->queryGenerator->generateQueryBuilder(
-            $this->builderPreparer,
-            $conditions,
-            $sortMethods,
-            $offset,
-            $limit
-        );
+        $queryBuilder = $this->entityManager->createQueryBuilder();
+        $this->builderPreparer->setWhereExpressions($conditions);
+        $this->builderPreparer->setOrderByExpressions($sortMethods);
+        $this->builderPreparer->fillQueryBuilder($queryBuilder);
+
+        // add offset if needed
+        if (0 !== $offset) {
+            $queryBuilder->setFirstResult($offset);
+        }
+
+        // add limit if needed
+        if (null !== $limit) {
+            $queryBuilder->setMaxResults($limit);
+        }
+
+        return $queryBuilder;
     }
 }